Audioplayer succesfully fetches from streamer. Still some issues with the controls (play/pause dynamic switch/hiding + volume control. To debug later)
