+++ HyperRadioMVC/Controllers/HomeController.cs
@@ -1,6 +1,7 @@
 using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using HyperRadioMVC.Models;
+using HyperRadioMVC.ViewModels;
 
 namespace HyperRadioMVC.Controllers;
 
@@ -15,15 +16,24 @@ public class HomeController : Controller
 
     public IActionResult Index()
     {
-        var artist = new ArtistProfile
+        
+        var viewModel = new HomeVM
         {
-            Name = "Frank Ocean",
-            Description = "Frank Ocean is an American singer, songwriter, record producer, and photographer. Known for his unconventional production and eclectic musical style, he has been acclaimed as one of the most influential contemporary artists.",
-            ImageUrl = "https://www.nme.com/wp-content/uploads/2016/09/2011FrankOcean02PR070212.jpg"
+            ArtistProfile = new ArtistProfileVM
+            {
+                Name = "Frank Ocean",
+                Description = "Frank Ocean is an American singer, songwriter, record producer, and photographer. Known for his unconventional production and eclectic musical style, he has been acclaimed as one of the most influential contemporary artists.",
+                ImageUrl = "https://www.nme.com/wp-content/uploads/2016/09/2011FrankOcean02PR070212.jpg"
+            },
+            ShowDetails = new ShowDetailsVM
+            {
+                Name = "Hyper Radio Live",
+                Description = "Weekly electronic music broadcast.",
+                ScheduledStart = DateTime.Now.AddHours(1)
+            }
         };
 
-        return View(artist);
-    }
+        return View(viewModel);    }
 
     public IActionResult Privacy()
     {
