--- HyperRadioMVC/ViewModels/ArtistProfileVM.cs
@@ -1,10 +1,0 @@
-namespace HyperRadioMVC.ViewModels;
-
-public class ArtistProfileVM
-{
-
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public string ImageUrl { get; set; }
-
-}
