--- HyperRadioMVC/ViewModels/ShowDetailsVM.cs
@@ -1,9 +1,0 @@
-namespace HyperRadioMVC.ViewModels;
-
-public class ShowDetailsVM
-{
-    public string Name { get; set; }
-    public string Description { get; set; } 
-    public DateTime ScheduledStart { get; set; }
-
-}
