+++ HyperRadioMVC/Views/Shared/_ArtistProfile.cshtml
@@ -1,4 +1,4 @@
-@model HyperRadioMVC.Models.ArtistProfile
+@model HyperRadioMVC.ViewModels.ArtistProfileVM
 
 
 <link rel="stylesheet" href="~/css/artistprofile.css" asp-append-version="true" />
