+++ HyperRadioMVC/Views/Shared/_ShowDetails.cshtml
@@ -1,4 +1,11 @@
-@model ShowDetailsVM
+@model HyperRadioMVC.ViewModels.ShowDetailsVM
+
+<head>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" href="~/css/showdetails.css" />
+</head>
+
+
 
 @{
 ViewBag.Title = "About Show";
