+++ HyperRadioMVC/bin/Debug/net8.0/HyperRadioMVC.staticwebassets.runtime.json
@@ -1,1 +1,1 @@
-{"ContentRoots":["/Users/joshuang/Desktop/Programming/Hyper Radio/Hyper-Radio-MVC/HyperRadioMVC/wwwroot/"],"Root":{"Children":{"css":{"Children":{"artistprofile.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/artistprofile.css"},"Patterns":null},"audioplayer.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/audioplayer.css"},"Patterns":null},"layout.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/layout.css"},"Patterns":null},"trackpitch.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/trackpitch.css"},"Patterns":null}},"Asset":null,"Patterns":null},"favicon.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"favicon.svg"},"Patterns":null},"js":{"Children":{"player.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"js/player.js"},"Patterns":null},"site.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"js/site.js"},"Patterns":null}},"Asset":null,"Patterns":null},"lib":{"Children":{"jquery-validation-unobtrusive":{"Children":{"jquery.validate.unobtrusive.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js"},"Patterns":null},"jquery.validate.unobtrusive.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"},"Patterns":null},"LICENSE.txt":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/LICENSE.txt"},"Patterns":null}},"Asset":null,"Patterns":null},"jquery-validation":{"Children":{"dist":{"Children":{"additional-methods.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/additional-methods.js"},"Patterns":null},"additional-methods.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/additional-methods.min.js"},"Patterns":null},"jquery.validate.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/jquery.validate.js"},"Patterns":null},"jquery.validate.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/jquery.validate.min.js"},"Patterns":null}},"Asset":null,"Patterns":null},"LICENSE.md":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/LICENSE.md"},"Patterns":null}},"Asset":null,"Patterns":null},"jquery":{"Children":{"dist":{"Children":{"jquery.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.js"},"Patterns":null},"jquery.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.min.js"},"Patterns":null},"jquery.min.map":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.min.map"},"Patterns":null}},"Asset":null,"Patterns":null},"LICENSE.txt":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/LICENSE.txt"},"Patterns":null}},"Asset":null,"Patterns":null}},"Asset":null,"Patterns":null},"logo.png":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.png"},"Patterns":null},"logo.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.svg"},"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
+{"ContentRoots":["/Users/joshuang/Desktop/Programming/Hyper Radio/Hyper-Radio-MVC/HyperRadioMVC/wwwroot/"],"Root":{"Children":{"css":{"Children":{"artistprofile.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/artistprofile.css"},"Patterns":null},"audioplayer.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/audioplayer.css"},"Patterns":null},"home.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/home.css"},"Patterns":null},"layout.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/layout.css"},"Patterns":null},"showdetails.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/showdetails.css"},"Patterns":null},"trackpitch.css":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"css/trackpitch.css"},"Patterns":null}},"Asset":null,"Patterns":null},"favicon.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"favicon.svg"},"Patterns":null},"js":{"Children":{"player.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"js/player.js"},"Patterns":null},"site.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"js/site.js"},"Patterns":null}},"Asset":null,"Patterns":null},"lib":{"Children":{"jquery-validation-unobtrusive":{"Children":{"jquery.validate.unobtrusive.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js"},"Patterns":null},"jquery.validate.unobtrusive.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"},"Patterns":null},"LICENSE.txt":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation-unobtrusive/LICENSE.txt"},"Patterns":null}},"Asset":null,"Patterns":null},"jquery-validation":{"Children":{"dist":{"Children":{"additional-methods.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/additional-methods.js"},"Patterns":null},"additional-methods.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/additional-methods.min.js"},"Patterns":null},"jquery.validate.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/jquery.validate.js"},"Patterns":null},"jquery.validate.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/dist/jquery.validate.min.js"},"Patterns":null}},"Asset":null,"Patterns":null},"LICENSE.md":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery-validation/LICENSE.md"},"Patterns":null}},"Asset":null,"Patterns":null},"jquery":{"Children":{"dist":{"Children":{"jquery.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.js"},"Patterns":null},"jquery.min.js":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.min.js"},"Patterns":null},"jquery.min.map":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/dist/jquery.min.map"},"Patterns":null}},"Asset":null,"Patterns":null},"LICENSE.txt":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"lib/jquery/LICENSE.txt"},"Patterns":null}},"Asset":null,"Patterns":null}},"Asset":null,"Patterns":null},"logo.png":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.png"},"Patterns":null},"logo.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.svg"},"Patterns":null},"pause.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"pause.svg"},"Patterns":null},"play.svg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"play.svg"},"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
