+++ HyperRadioMVC/wwwroot/css/artistprofile.css
@@ -1,4 +1,3 @@
-/* Fill the main content completely */
 .artist-profile-container {
     flex: 1 1 auto;  
     display: flex;
@@ -10,7 +9,6 @@
     box-sizing: border-box;
 }
 
-/* Image container */
 .artist-image {
     position: relative;
     flex: 1 1 400px;
@@ -25,9 +23,8 @@
     border-radius: 10px;
 }
 
-/* Name overlay */
 .artist-name {
-    position: absolute;
+    position: relative;
     bottom: 20px;
     left: 50%;
     transform: translateX(-50%);
@@ -39,7 +36,6 @@
     color: white;
 }
 
-/* Description box */
 .artist-description {
     flex: 1 1 200px;
     max-width: 400px;
@@ -50,7 +46,6 @@
     line-height: 1.6;
 }
 
-/* Responsive adjustments */
 @media (max-width: 900px) {
     .artist-profile-container {
         flex-direction: column;
