+++ HyperRadioMVC/wwwroot/css/home.css
@@ -1,0 +1,42 @@
+.split-view {
+    display: flex;
+    height: 100vh; /* full height */
+    background-color: #000;
+    color: #fff;
+    overflow: hidden;
+}
+
+.left-panel {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 50%;
+    height: 100vh;
+    background-color: black ;
+    padding: 1.5rem;
+    overflow-y: hidden;
+}
+
+.right-panel {
+    margin-left: 50%; /* start where the left panel ends */
+    width: 50%;
+    height: 100vh;
+    padding: 2rem;
+    overflow-y: auto;
+    background-color: #000;
+}
+
+/* Scrollbar styling */
+.right-panel::-webkit-scrollbar,
+.left-panel::-webkit-scrollbar {
+    width: 8px;
+}
+.right-panel::-webkit-scrollbar-thumb,
+.left-panel::-webkit-scrollbar-thumb {
+    background: #555;
+    border-radius: 4px;
+}
+.right-panel::-webkit-scrollbar-thumb:hover,
+.left-panel::-webkit-scrollbar-thumb:hover {
+    background: #777;
+}
