+++ HyperRadioMVC/wwwroot/css/showdetails.css
@@ -1,0 +1,15 @@
+.container {
+    display: block;
+    justify-content: space-around;
+    align-items: center;
+    background-color: red;
+    padding: 30px;
+    border-radius: 45px;
+    margin-right: 30%;
+    margin-left: 10%;
+    margin-top: 50%;
+}
+
+.show-time {
+    font-size: 30px;
+}
